requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
PyGithub>=2.1.1
weaviate-client>=4.5.0
docker>=7.0.0
playwright>=1.43.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
requests>=2.31.0
python-dotenv>=1.0.0
